final/back/requirements/base.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.5
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary>=2.9.9
djangorestframework-simplejwt==5.5.0
pyjwt==2.9.0
markdown==3.7
django-filter==25.1
drf-spectacular==0.28.0
boto3==1.37.33
requests==2.32.3
django-cors-headers==4.7.0